Direct new repos to use an initial PR
This update is in response to discussion here: https://github.com/cisagov/development-guide/pull/44#discussion_r558366304

We want to direct users to make an initial branch and PR with their code instead of directly committing to `develop` before enabling branch protection.

Also adds a table of contents to the project setup guide and clarifies a point about status checks in the branch protection guide.