Add a package_dir attribute
Since this Python module doesn't use a standard src-layout, setuptools
gets confused and is unable to locate the packages to be included.
Since this "Python module" is really just a collection of scripts, we
can tell setuptools to restrict itself to project_setup/scripts.

This is yet another reason why #33 is a good idea.