Remove references to Trio
Update these references to point to CISA GWE instead, where
appropriate.

Also remove references to JIRA and the dev team GH project board,
since the former is going away and the latter is of limited use.

Glom together sections on dev and non-dev team members.  This document
is for dev team members and QA team members only, so it doesn't make
sense to break out a separate section for non-members.