Remove plugin configs and move instructions
It was discovered that the docker-compose formula is the Compose plugin.
It creates a docker-compose binary that is symlinked into the plugins
directory. This is also the case for the docker-buildx plugin. Since this
is the case it makes sense to remove the instructions for configuring the
plugins.

The Colima instructions were also moved to be a sub section under
the Docker section since Colima relies on Docker under the hood and is
simply an alternative to Docker.